Final_project/kafka/requirements.txt
aiofiles==23.2.1
aiogram==2.21
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
confluent-kafka==2.3.0
exceptiongroup==1.2.0
fastapi==0.104.1
Flask==3.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-final-package==0.1.4
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
pdfkit==1.0.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pythonnet==3.0.3
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
yarl==1.9.4
